requirements/requirements.txt
Django==2.0.2
django-cors-headers==2.1.0
django-environ==0.4.4
django-filter==1.1.0
django-graphql-jwt==0.1.4
django-model-utils==3.1.1
graphene==2.0.1
graphene-django==2.0.0
graphql-core==2.0
graphql-relay==0.4.5
iso8601==0.1.12
promise==2.1
PyJWT==1.5.3
pytz==2017.3
Rx==1.6.0
singledispatch==3.4.0.3
six==1.11.0
typing==3.6.4
